Merge branch tauri-2 into main

See merge request 1io/co-sdk!6

+++ co-messaging/src/matrix_event/call_event.rs
@@ -1,10 +1,11 @@
 use crate::{EventContent, EventType};
+use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 /**
  * Session description object for sdp offers and answers
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct SessionDescription {
 	pub sdp: String,
 	#[serde(rename = "type")]
@@ -20,17 +21,17 @@ impl SessionDescription {
 /**
  * ICE candidate for WebRTC exchange protocol
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct ICECandidate {
 	pub candidate: String, // SDP 'a' line of the candidate
 	#[serde(rename = "sdpMLineIndex")]
-	pub sdp_m_line_index: i64, // index of the SDP 'm' line this candidate is intended for
+	pub sdp_m_line_index: u32, // index of the SDP 'm' line this candidate is intended for
 	#[serde(rename = "sdpMid")]
 	pub sdp_m_id: String, // the SDP media type this candidate is intended for
 }
 
 impl ICECandidate {
-	pub fn new(candidate: impl Into<String>, sdp_m_id: impl Into<String>, sdp_m_line_index: i64) -> Self {
+	pub fn new(candidate: impl Into<String>, sdp_m_id: impl Into<String>, sdp_m_line_index: u32) -> Self {
 		Self { candidate: candidate.into(), sdp_m_id: sdp_m_id.into(), sdp_m_line_index }
 	}
 }
@@ -42,22 +43,22 @@ impl ICECandidate {
  * across all participants. version: The version of the VoIP specs used for the message. This version is "1". A
  * string is used for experimental versions.
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 #[serde(tag = "type", content = "content")]
 pub enum CallType {
-	#[serde(rename = "m.call.invite")]
+	#[serde(rename = "call_invite")]
 	Invite(CallInviteContent),
-	#[serde(rename = "m.call.answer")]
+	#[serde(rename = "call_answer")]
 	Answer(AnswerCallContent),
-	#[serde(rename = "m.call.candidates")]
+	#[serde(rename = "call_candidates")]
 	Candidates(CallCandidatesContent),
-	#[serde(rename = "m.call.select_answer")]
+	#[serde(rename = "call_select_answer")]
 	SelectAnswer(SelectCallAnswerContent),
-	#[serde(rename = "m.call.negotiate")]
+	#[serde(rename = "call_negotiate")]
 	Negotioation(CallNegotiationContent),
-	#[serde(rename = "m.call.reject")]
+	#[serde(rename = "call_reject")]
 	Reject(RejectCallContent),
-	#[serde(rename = "m.call.hangup")]
+	#[serde(rename = "call_hangup")]
 	Hangup(HangupCallContent),
 }
 
@@ -84,14 +85,14 @@ impl From<CallType> for EventContent {
 /**
  * Initial event to invite other parties to a call
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct CallInviteContent {
 	pub call_id: String,
 	pub party_id: String,
 	pub version: String,
 	#[serde(skip_serializing_if = "Option::is_none")]
 	pub invitee: Option<String>, // DID of the called user. Any user in room may answer if omitted
-	pub lifetime: i64,             // Time in ms during which invite is valid after sending this event
+	pub lifetime: u32,             // Time in ms during which invite is valid after sending this event
 	pub offer: SessionDescription, // Session description object
 }
 
@@ -112,7 +113,7 @@ impl CallInviteContent {
 		call_id: impl Into<String>,
 		party_id: impl Into<String>,
 		invitee: Option<String>,
-		lifetime: i64,
+		lifetime: u32,
 		offer_sdp: impl Into<String>,
 	) -> Self {
 		Self {
@@ -129,7 +130,7 @@ impl CallInviteContent {
 /**
  * Event used when answering an invite event
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct AnswerCallContent {
 	pub call_id: String,
 	pub party_id: String,
@@ -163,7 +164,7 @@ impl AnswerCallContent {
 /**
  * Event used to exchange viable ICE candidates with the other party upon answering a call
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct CallCandidatesContent {
 	pub call_id: String,
 	pub party_id: String,
@@ -192,7 +193,7 @@ impl CallCandidatesContent {
 /**
  * Event used to select one of possibly multiple call answers
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct SelectCallAnswerContent {
 	pub call_id: String,
 	pub party_id: String,
@@ -228,7 +229,7 @@ impl SelectCallAnswerContent {
  * then send an answer. Offer and answer should never both be set. To ensure this they are not public to force the
  * users to use the setters.
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct CallNegotiationContent {
 	pub call_id: String,
 	pub party_id: String,
@@ -238,7 +239,7 @@ pub struct CallNegotiationContent {
 	#[serde(skip_serializing_if = "Option::is_none")]
 	offer: Option<SessionDescription>, // Session description object for negotioation offers
 	#[serde(skip_serializing_if = "Option::is_none")]
-	lifetime: Option<i64>, // Time in ms before offer timeout
+	lifetime: Option<u32>, // Time in ms before offer timeout
 }
 
 impl From<CallNegotiationContent> for EventContent {
@@ -258,7 +259,7 @@ impl CallNegotiationContent {
 		call_id: impl Into<String>,
 		party_id: impl Into<String>,
 		offer_sdp: impl Into<String>,
-		lifetime: i64,
+		lifetime: u32,
 	) -> Self {
 		Self {
 			call_id: call_id.into(),
@@ -279,7 +280,7 @@ impl CallNegotiationContent {
 			lifetime: None,
 		}
 	}
-	pub fn set_offer(&mut self, offer_sdp: impl Into<String>, lifetime: i64) {
+	pub fn set_offer(&mut self, offer_sdp: impl Into<String>, lifetime: u32) {
 		self.offer = Some(SessionDescription::new(offer_sdp, "offer"));
 		self.lifetime = Some(lifetime);
 		self.answer = None;
@@ -295,7 +296,7 @@ impl CallNegotiationContent {
 	pub fn get_answer(&self) -> Option<SessionDescription> {
 		self.answer.clone()
 	}
-	pub fn get_lifetime(&self) -> Option<i64> {
+	pub fn get_lifetime(&self) -> Option<u32> {
 		self.lifetime
 	}
 }
@@ -303,7 +304,7 @@ impl CallNegotiationContent {
 /**
  * Event sent if call was rejected by a user.
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct RejectCallContent {
 	pub call_id: String,
 	pub party_id: String,
@@ -331,7 +332,7 @@ impl RejectCallContent {
 /**
  * Enum containg possible reasons for a hangup event
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub enum HangupCallReason {
 	#[serde(rename = "ice_failed")]
 	IceFailed, // ICE negotiation has failed and connection could not be established
@@ -354,7 +355,7 @@ pub enum HangupCallReason {
 /**
  * Hangup event used to signal the termination of the call.
  */
-#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone, JsonSchema)]
 pub struct HangupCallContent {
 	pub call_id: String,
 	pub party_id: String,
